requirement.txt
pandas
tenacity
openai
streamlit
python-dotenv
watchdog